Move PubSub and Presence to separate components
